pyproject.toml
[project]
name = "pumping-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
